Skip dropping tablesync replication slots on the publisher (#593)
when we're in the process of dropping subscriptions
inherited by a neon branch.
Because these slots are still needed by the parent branch subscriptions.

For regular slots we handle this by setting the slot_name to NONE before
calling DROP SUBSCRIPTION, but tablesync slots are not exposed to SQL.

rely on GUC disable_logical_replication_subscribers=true
to know that we're in the Neon-specific process of dropping subscriptions.