Add hook in pg_signal_backend()
The purpose of the hook is to facilitate Lakebase's behavior of not
allowing customers to cancel queries on replicas.

Lakebase exposes a single endpoint to customers unlike Neon, where there
is an endpoint for the primary and all replicas. Lakebase has had
customer issues where they do the following:

	# On connection 1
	SELECT pid from pg_stat_activity WHERE ...;

	# On connection 2
	SELECT pg_cancel_backend(pid) -- Where pid is from connection 1

Connection 2 is not guaranteed to go to the same endpoint, so a customer
may call pg_cancel_backend() on the wrong endpoint. Commonly, this will
end up in an error because the PID does not exist, but in rare cases
PIDs will align across endpoints, and the customer will cancel a query
they didn't want to cancel.

Whether we want to keep this hook indefinitely requires talking to
product, and explaining the situation.

Signed-off-by: Tristan Partin <tristan.partin@databricks.com>