Change how sorted GiST index build is handled with Neon storage in v14-v16
Reorder writing the page with smgrextend() and WAL-logging it, so that
the page is WAL-logged first. That makes it work the same it
effectively does on v17, where the we use the new bulk-writing
facility. That's also how B-tree index build does it.

This allows us to reinstate the check that an unlogged index build is
only performed on a newly-created, i.e. empty, relation, and remove
the hack to invalidate the bogus page from the LFC at the end of an
unlogged index build. That was added just to handle the special way
that sorted GiST index build abused the unlogged index build
machinery.